--- Merging r72037 into '.': U    lib/Target/ARM/ARMISelLowering.cpp
Fix pr4202: Disable CodePlacementOpt for ARM.  The ARMConstantIslandPass has
to run last because it needs to know the exact size and position of every
basic block.  Currently CodePlacementOpt is set up to run last.  It might be
worthwhile to investigate reordering these passes, but for now, let's just
make it work.