--- Merging r74056 into '.': U    gcc/llvm-convert.cpp
Fix an array over-read problem that would sometimes manifest as a crash (when overreading
a string would point to invalid data.  The issue here is that TREE_STRING_LENGTH returns
the length in bytes of a string, not in characters.  This is different for a wide string.

In cases when this didn't crash, this bug would not manifest as a codegen problem, because
the code right below the overread would trim the length of the generated vector to the 
right length.

This fixes rdar://6961178